pyproject.toml
[tool.poetry]
name = "registry-viewer"
version = "0.1.0"
description = ""
authors = ["Roman Slyusar"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.8.10"
django = "^3.2.18"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
tabulate = "^0.9.0"
python-dotenv = "^1.0.0"
pylightxl = "^1.59"
tqdm = "^4.64.0"
psutil = "^5.9.4"
schedule = "^1.1.0"
chardet = "^5.1.0"
pyuac = "0.0.3"
pypiwin32 = "^223"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
